pyproject.toml
[tool.poetry]
name = "exquisite_cybertree"
version = "0.1.0"
description = ""
authors = ["Rick Anderson <rick.rickanderson@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.21.1"
matplotlib = "^3.6.3"
numpy = "^1.24.2"
moviepy = "^1.0.3"
openai = "^1.33.0"
streamlit = "^1.35.0"
pillow = "9.5.0"
ffmpeg-python = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
